Revert "feat(sdk-core): extract canvas components from canvas-r3f (task sdk-3)"
