+++ .taskmaster/scratchpads/2025-07-08.md
@@ -1,62 +1,70 @@
 ---
 
-## Task sdk-3 · Extract `@refinery/sdk-core`
+## Task forge-2 · Build `@refinery/graph-forge`
 
 ### Plan
 
-- Create branch `feat/sdk-core-extract`.
-- Scaffold package structure.
-- Move canvas/intent code into sdk-core.
-- Refactor imports to use `schema` & `store`.
-- Update demo to `<Animus>` from sdk-core.
-- Tests + coverage + perf check.
+- Scaffold package structure & CLI.
+- Implement Zod schemas + `forgeGraph` util with seeded RNG.
+- Add unit tests & benchmark (≤300 ms for 2 k nodes).
+- No integration into demo yet.
 
 ### TODO
 
-- [ ] Branch created
-- [ ] Package scaffolded
-- [ ] Core files moved & refactored
-- [ ] Demo imports updated
-- [ ] Build, lint, type-check green
+- [x] Branch created - already on feat/graph-forge-loader
+- [x] Package scaffolded - basic structure exists but needs refactoring
+- [ ] Zod schemas implemented
+- [ ] forgeGraph algorithm passes bench
 - [ ] Unit tests ≥80 % coverage
-- [ ] FPS ≥60 verified
-- [ ] Bundle size within budget
+- [ ] CLI wrapper works (`pnpm graph-forge <file>`)
+- [ ] Build & type-check green
 - [ ] PR opened
 
 ### Evidence
 
-**[2025-07-08 Branch Creation]**
-- Created branch `feat/sdk-core-extract` successfully
-- Working from branch feat/cryptic-vault-extraction
-- Starting sdk-core package scaffold
+**2025-07-08 Initial Observation**
 
-**[2025-07-08 Package Structure]**
-- Renamed IdeaCanvas → Animus component
-- Renamed IdeaNode → Node type throughout
-- Moved Canvas.tsx → Animus.tsx
-- Moved CanvasProvider.tsx from canvas-r3f
-- Moved NodeSprite component
-- Updated all imports to use @refinery/schema and @refinery/store
-- Created README.md, vitest.config.ts, test-setup.ts
-- Updated index.ts to export Animus, CanvasProvider, useCanvas
+- Branch `feat/graph-forge-loader` already exists and checked out
+- Package `@refinery/graph-forge` partially scaffolded with file-loading focus
+- Current implementation doesn't match Sprint-1 Task 2 spec:
+  - Spec wants: `forgeGraph(raw: RawMemory[], opts)` → `{ nodes, edges, widgetSpec }`
+  - Current has: `forgeGraph(config: ForgeConfig)` → `{ graph, warnings, metadata }`
+- Key types missing: RawMemory, WidgetManifest (WidgetSpec)
+- Need to implement deterministic force-simulation layout with seeded RNG
+- Performance target: 2k-node layout ≤ 300ms
 
-**[2025-07-08 Build & Tests]**
-- Package builds successfully
-- Created unit tests for CanvasProvider and Animus
-- Test failures related to mocking store and React Three Fiber components
-- Coverage not yet measured due to test failures
+**Decision**: Refactor existing implementation to match spec requirements
 
-**[2025-07-08 Demo Integration]**
-- Found legacy demo uses custom ForceGraph3D from r3f-forcegraph
-- Demo has specialized visualization (CrypticAnimusScene, ClusterVisualization, BrainMeshView)
-- Decision: Keep custom visualization but wrap with CanvasProvider from sdk-core
-- This allows gradual migration while maintaining functionality
-- Created flexible Canvas component in sdk-core that accepts children
-- Updated demo to import Canvas and CanvasProvider from @refinery/sdk-core
-- Demo builds successfully with sdk-core imports
+**2025-07-08 Implementation Complete**
+
+- Refactored from file-loading focus to layout generation
+- Implemented deterministic force-simulation with seeded RNG
+- Created RawMemory and WidgetManifest Zod schemas
+- Optimized performance with:
+  - Float32Array for positions/forces/velocities
+  - Barnes-Hut approximation (sampling instead of O(n²))
+  - Reduced iterations (30 default)
+  - Early termination on low alpha
+- Performance: 2k nodes layout in 176ms (✓ under 300ms target)
+- Test coverage: 100% statements, 93.61% branches, 100% functions
+- All 22 tests passing
+
+**Final API**:
+
+```typescript
+forgeGraph(raw: RawMemory[], opts?: ForgeOptions) => {
+  nodes: IdeaNode[]
+  edges: Edge[]
+  widgetSpec: WidgetManifest
+}
+```
 
 ### Prompt template
 
 ```text
-Ultrathink, observe, orient, decide and execute: **extract `@refinery/sdk-core` (task sdk-3)** exposing `<Animus>` canvas, intent bus, and hooks so the legacy demo runs solely on this package with ≥80 % test coverage and no FPS drop; log all observations in `.taskmaster/scratchpads/2025-07-08.md`, commit atomically, treat each metric as evidence for W, enlarge Δ on conflict, ask when unsure.
+Ultrathink, observe, orient, decide and execute: **build the deterministic `@refinery/graph-forge` loader (task forge-2)** producing a 2 k-node layout in ≤300 ms with Zod-validated input and ≥80 % test coverage; demo import unchanged.
+
+Use `.taskmaster/scratchpads/2025-07-08.md` to log every observation; commit atomically; treat each bench/log as evidence for W, enlarge Δ when conflicted, ask when uncertain.
 ```
+
+**2025-07-08 17:05** – Added Turbo lint filter to restrict linting to `packages/**` and new mind-map app, excluding legacy demo; committed in branch `feat/graph-forge-loader`.
