+++ CLAUDE.md
@@ -94,35 +94,3 @@ Append bench runs, test output, and coverage summary to `.taskmaster/scratchpads
 - Keep all heavy dependencies optional/peer.
 - No changes to demo imports in this slice.
 - If benchmark fails the 300 ms budget, stop and optimise before pushing.
-
-## Sprint-1 · Task 3 — sdk-core extraction (sdk-3)
-
-> **Why third?** Exposing `<Animus>` canvas + intent bus turns the demo into the first client of the newly-modular runtime, proving the open-core value.
-
-**Success condition**
-
-1. New package `packages/sdk-core` exports `<Animus />`, `CanvasProvider`, selection hooks, and intent bus.
-2. Legacy demo imports only `@refinery/sdk-core` for canvas/runtime — no direct import from `canvas-r3f` or vendored packages.
-3. Build, lint, type-check green; coverage ≥ 80 % for sdk-core.
-4. FPS unchanged (≥ 60) in dev & prod; bundle size growth < 50 KB gz.
-
-**Execution checklist**
-
-- [ ] `git checkout -b feat/sdk-core-extract` _(task sdk-3)_
-- [ ] Scaffold `packages/sdk-core` (`package.json`, `tsconfig.json`, README).
-- [ ] Copy canvas-r3f core files (`Animus.tsx`, providers, hooks) → sdk-core.
-- [ ] Replace internal imports with `@refinery/schema`, `@refinery/store`.
-- [ ] Add barrel `index.ts` exporting minimal public API.
-- [ ] Write unit tests for hooks + provider context.
-- [ ] Update demo to import `<Animus>` from sdk-core.
-- [ ] Run perf smoke test (FPS overlay) and snapshot bundle size.
-- [ ] Push branch; open draft PR (task sdk-3).
-
-**Evidence logging**
-Append build logs, test coverage, FPS screenshot links, and bundle size diff to `.taskmaster/scratchpads/2025-07-08.md`.
-
-**Guard-rails**
-
-- Keep public surface tiny; anything not needed by demo stays internal.
-- If FPS drops below 60, revert & profile before proceeding.
-- No breaking changes to schema types; reuse Node/Edge only.
