--- packages/sdk-core/src/components/index.ts
@@ -1,1 +1,0 @@
-export { NodeSprite } from './NodeSprite'
