+++ packages/sdk-core/src/index.ts
@@ -1,13 +1,10 @@
-// Main canvas component (Animus)
-export { Animus } from './Animus'
-export type { AnimusProps } from './Animus'
+// Scene component for rendering 3D graph content
+export { Scene } from './Scene'
+export type { SceneProps } from './Scene'
 
-// Canvas component for custom content
-export { Canvas } from './Canvas'
-export type { CanvasProps } from './Canvas'
-
-// Canvas provider and hooks
-export { CanvasProvider, useCanvas } from './CanvasProvider'
+// IdeaCanvas component for the main canvas interface
+export { IdeaCanvas } from './IdeaCanvas'
+export type { IdeaCanvasProps } from './IdeaCanvas'
 
 // Intent bus for command processing and state management
 export { 
@@ -17,5 +14,5 @@ export {
 export type { CanvasState } from './IntentBus'
 
 // Re-export types from dependencies for convenience
-export type { Node, Edge } from '@refinery/schema'
+export type { IdeaNode, Edge } from '@refinery/schema'
 export type { RendererCommand } from '@refinery/store'
