+++ packages/sdk-core/tsconfig.json
@@ -11,7 +11,6 @@
     "tsBuildInfoFile": "./tsconfig.tsbuildinfo"
   },
   "include": ["src/**/*.ts", "src/**/*.tsx"],
-  "exclude": ["src/**/*.test.ts", "src/**/*.test.tsx", "test-setup.ts"],
   "references": [
     { "path": "../schema" },
     { "path": "../store" }
